First commit of a BF interpreter.
I mainly want to commit this to save the code I have right now.
But I'm basically going to redesign it as follows:

Read the file character by character via a file strea (fgetc).
There are two modes: immediate mode and loop mode. In immediate mode,
characters are simply executed (e.g., '+' causes ++*mem and so on.)
They are, in other words, executed immediately, then discarded (hence
the term "immediate mode".)

If a '[' is encountered, then the program enters "loop mode":
the loop body (everything up to just before the next ']') is loaded
into a buffer. If EOF is encountered during this process, then the
program aborts with an error message; this is how unbalanced
loop-brackets are detected. The contents of the loop buffer are
executed as long as *mem != 0, upon which the loop buffer
is reset, and control returns to immediate mode.